Initial Framework for server integration
Put in the updates to include nodejs server side and reorganized the
client side within the client folder (probably should have called it
public to be consistent with the web but it felt conflicting with the
server side generated content so I changed it).

Also incorporated the proper use of bower for the UI modules. I was
unable to properly create the angular-ui-bootstrap bower module however
so that bit is still manually done (could not get the bower module to
compile with the templates included in with the glyphicons changed to
fa in order to support font awesome icons instead of the bootstrap
icons.