More reorganization to better fit developers
The client folder is now the final artifact location for the UI so any
IDE that is able to utilize css/js references (pretty much anything
except for eclipse as far as I can tell) will now function correctly.

The custom client side js and less files are moved to top level of the
project and compiled into the client folder condensed.

Also started separating out the reusable modules portion of angular
code from the seed framework code for the application itself.

Made it easier to find and start the "main" server by calling the file
server.js and placing it at the top level of the project. All other
server side code still resides within the server folder however.

Also added support for all 404/500 errors to be returned in json form
always. This will later be enhanced to return everything in the unified
error response format once that is wrapped up.